[WonIOne] 수강 후기 평점 데이터 타입 변경 author: 김윤환 file: 	BoardController.java 	addReview.html 	Review.java detail: 변경내용 	수강 후기 평점 데이터 타입 변경 	int -> double 	평점 소수 선택 가능

+++ wonione/src/main/java/cafe24/wio/controller/BoardController.java
@@ -243,7 +243,7 @@ public class BoardController {
 			  					,Model model
 			  					,@RequestParam(value="reviewNum", required = false) String reviewNum
 			  					,@RequestParam(value="reviewWriterId", required = false) String reviewWriterId
-			  					,@RequestParam(value="reviewGrade", required = false) int reviewGrade
+			  					,@RequestParam(value="reviewGrade", required = false) double reviewGrade
 			  					,@RequestParam(value="reviewRegDate", required = false) String reviewRegDate
 			  					,@RequestParam(value="reviewComment", required = false) String reviewComment	  					
 			  					) {
